+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/update/CharacterUpdateTask.kt
@@ -1,7 +1,9 @@
-package world.gregs.voidps.engine.client.update.task
+package world.gregs.voidps.engine.client.update
 
-import world.gregs.voidps.engine.client.update.task.npc.NPCUpdateTask
-import world.gregs.voidps.engine.client.update.task.player.PlayerUpdateTask
+import world.gregs.voidps.engine.client.update.batch.ChunkBatches
+import world.gregs.voidps.engine.client.update.iterator.TaskIterator
+import world.gregs.voidps.engine.client.update.npc.NPCUpdateTask
+import world.gregs.voidps.engine.client.update.player.PlayerUpdateTask
 import world.gregs.voidps.engine.entity.character.CharacterList
 import world.gregs.voidps.engine.entity.character.player.Player
