+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/update/MovementTask.kt
@@ -1,5 +1,6 @@
-package world.gregs.voidps.engine.client.update.task
+package world.gregs.voidps.engine.client.update
 
+import world.gregs.voidps.engine.client.update.iterator.TaskIterator
 import world.gregs.voidps.engine.entity.Direction
 import world.gregs.voidps.engine.entity.character.*
 import world.gregs.voidps.engine.entity.character.move.moving
