+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/update/batch/ChunkExtensions.kt
@@ -1,4 +1,4 @@
-package world.gregs.voidps.engine.client.update.task
+package world.gregs.voidps.engine.client.update.batch
 
 import world.gregs.voidps.engine.entity.definition.AnimationDefinitions
 import world.gregs.voidps.engine.entity.gfx.AreaGraphic
