+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/update/iterator/ParallelIterator.kt
@@ -1,7 +1,8 @@
-package world.gregs.voidps.engine.client.update.task
+package world.gregs.voidps.engine.client.update.iterator
 
 import com.github.michaelbull.logging.InlineLogger
 import it.unimi.dsi.fastutil.objects.ObjectArrayFIFOQueue
+import world.gregs.voidps.engine.client.update.CharacterTask
 import world.gregs.voidps.engine.entity.character.Character
 import java.util.concurrent.Executors
 import java.util.concurrent.Future
