+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/update/iterator/SequentialIterator.kt
@@ -1,5 +1,6 @@
-package world.gregs.voidps.engine.client.update.task
+package world.gregs.voidps.engine.client.update.iterator
 
+import world.gregs.voidps.engine.client.update.CharacterTask
 import world.gregs.voidps.engine.entity.character.Character
 
 class SequentialIterator<C : Character> : TaskIterator<C> {
