+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/update/iterator/TaskIterator.kt
@@ -1,5 +1,6 @@
-package world.gregs.voidps.engine.client.update.task
+package world.gregs.voidps.engine.client.update.iterator
 
+import world.gregs.voidps.engine.client.update.CharacterTask
 import world.gregs.voidps.engine.entity.character.Character
 
 interface TaskIterator<C : Character> {
