+++ engine/src/test/kotlin/world/gregs/voidps/engine/client/update/task/CharacterUpdateTaskTest.kt
@@ -6,8 +6,11 @@ import io.mockk.spyk
 import io.mockk.verify
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import world.gregs.voidps.engine.client.update.task.npc.NPCUpdateTask
-import world.gregs.voidps.engine.client.update.task.player.PlayerUpdateTask
+import world.gregs.voidps.engine.client.update.CharacterUpdateTask
+import world.gregs.voidps.engine.client.update.batch.ChunkBatches
+import world.gregs.voidps.engine.client.update.iterator.SequentialIterator
+import world.gregs.voidps.engine.client.update.npc.NPCUpdateTask
+import world.gregs.voidps.engine.client.update.player.PlayerUpdateTask
 import world.gregs.voidps.engine.entity.character.npc.NPCs
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.Players
