+++ engine/src/test/kotlin/world/gregs/voidps/engine/client/update/task/NPCUpdateTaskTest.kt
@@ -11,7 +11,7 @@ import world.gregs.voidps.buffer.read.BufferReader
 import world.gregs.voidps.buffer.write.BufferWriter
 import world.gregs.voidps.buffer.write.Writer
 import world.gregs.voidps.cache.definition.data.NPCDefinition
-import world.gregs.voidps.engine.client.update.task.npc.NPCUpdateTask
+import world.gregs.voidps.engine.client.update.npc.NPCUpdateTask
 import world.gregs.voidps.engine.entity.Direction
 import world.gregs.voidps.engine.entity.character.npc.NPC
 import world.gregs.voidps.engine.entity.character.npc.NPCs
