+++ engine/src/test/kotlin/world/gregs/voidps/engine/map/chunk/ChunkBatchesTest.kt
@@ -6,7 +6,7 @@ import io.mockk.mockkStatic
 import io.mockk.verify
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import world.gregs.voidps.engine.client.update.task.ChunkBatches
+import world.gregs.voidps.engine.client.update.batch.ChunkBatches
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.name
 import world.gregs.voidps.engine.event.eventModule
