+++ game/src/main/kotlin/world/gregs/voidps/world/map/BorderGuard.kts
@@ -4,7 +4,7 @@ import kotlinx.coroutines.NonCancellable
 import kotlinx.coroutines.withContext
 import world.gregs.voidps.engine.action.ActionType
 import world.gregs.voidps.engine.action.action
-import world.gregs.voidps.engine.client.update.task.animate
+import world.gregs.voidps.engine.client.update.batch.animate
 import world.gregs.voidps.engine.entity.*
 import world.gregs.voidps.engine.entity.character.Moved
 import world.gregs.voidps.engine.entity.character.Moving
